+++ Learnia/Learnia/src/main/java/com/proyecto/Learnia/service/CategoriaService.java
@@ -6,7 +6,7 @@ import java.util.List;
 public interface CategoriaService {
     List<Categoria> listar();
     Categoria crear(Categoria categoria);
-    Categoria actualizar(Integer id, Categoria categoria);
-    Categoria buscarPorId(Integer id);
-    void eliminar(Integer id);
+    Categoria actualizar(Long id, Categoria categoria);
+    Categoria buscarPorId(Long id);
+    void eliminar(Long id);
 }
